ScorePreloadAndPark: D.R.Y. up the code a little
With this new Base base, now the Red/Blue autonomous classes need only
Extend the base and set the variables that are different for themselves.
